# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TAPER",
        "version": "0.1"
    },
    "paths": {
...
        "/api/film-copy/current": {
            "get": {
                "tags": [
                    "Фильмокопии"
                ],
                "summary": "Получение списка баннеров",
                "operationId": "59f6a9ea8f6d085ba65abd7ab93e5b40",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/film-copy/future": {
            "get": {
                "tags": [
                    "Фильмокопии"
                ],
                "summary": "Получение списка фильмов предстоящих в прокате",
                "operationId": "5b901decc6488d2c30109e1cda0bf9eb",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/film-copy/ps": {
            "get": {
                "tags": [
                    "Фильмокопии"
                ],
                "summary": "Получение списка фильмов в киноклубе",
                "operationId": "0263e968487b1055fcab4ffb416cb2d8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/film-copy/retro": {
            "get": {
                "tags": [
                    "Фильмокопии"
                ],
                "summary": "Получение списка фильмов ретро",
                "operationId": "2842e75df07e0e179e2a50dbc9388cfc",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/film-copy/not-only-jazz": {
            "get": {
                "tags": [
                    "Фильмокопии"
                ],
                "summary": "Получение списка фильмов концертов",
                "operationId": "dbc26a59fa7df5cbc22f286f71676dfd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/film-copy/specific/{id}": {
            "get": {
                "tags": [
                    "Фильмокопии"
                ],
                "summary": "Получение фильма по id",
                "operationId": "708db9a81c25138257977ec5b26a3204",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/film-copy/movie-schedule": {
            "get": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Получение списка баннеров",
                "operationId": "50cbc24b74b5472665e138a7951fc6a9",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/film-copy/movie-schedule/{performance}": {
            "get": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Получение списка баннеров",
                "operationId": "0a081d1317fecea38e1ef3858c10fce4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/film-copy/status/{performance}": {
            "get": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Получение списка баннеров",
                "operationId": "ce6041f6f5e429699b0d9794bbcf157d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/film-copy/status-auth/{performance}": {
            "get": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Получение списка баннеров",
                "operationId": "0b51d1c8422c48fcdc40a15ddd6d1735",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}